chore: split out types correctly (#129) (#129)
* chore: split out styles

* chore: seperate styles for each component

* chore: remove duplicate files

* chore: split out types
