+++ src/modules/Projects/types.ts
@@ -22,3 +22,17 @@ export interface GetProjectsSuccessAction {
 }
 
 export type ProjectsActionTypes = GetProjectsAction | GetProjectsSuccessAction
+
+export interface Stats {
+  claims: Claims
+  totalServiceProviders: number
+  totalProjects: number
+  totalEvaluationAgents: number
+}
+interface Claims {
+  total: number
+  totalSuccessful: number
+  totalSubmitted: number
+  totalPending: number
+  totalRejected: number
+}
